added lots of function and top-level file comments
