+++ client/src/Authentication/RegisterPage.tsx
@@ -15,6 +15,10 @@ import ScreenGrid from '../components/ScreenGrid';
 import FormRow from '../components/form/FormRow';
 import FormGrid from '../components/form/FormGrid';
 
+/**
+ * A page users visit to be able to register for a new account by inputting
+ * fields such as their name, email, and password.
+ */
 function RegisterPage() {
   const navigate = useNavigate();
