+++ client/src/assets/types.ts
@@ -1,4 +1,8 @@
 /**
+ * A file for defining various types used throughout the project.
+ */
+
+/**
  * A type for encapsulating any react components/elements as children. The child
  * or children is accessed from the `children` property.
  */
