+++ server/src/models/session.ts
@@ -1,3 +1,7 @@
+/**
+ * Defines the Session model for the database, which stores information
+ * about user sessions, and also the interface to access the model in TypeScript.
+ */
 import mongoose from 'mongoose';
 
 interface ISession extends mongoose.Document {
