+++ server/src/routes/admin.route.ts
@@ -1,0 +1,29 @@
+/**
+ * Specifies the middleware and controller functions to call for each route
+ * relating to admin users.
+ */
+import express from 'express';
+import { isAdmin } from '../controllers/admin.middleware';
+import {
+  getAllUsers,
+  upgradePrivilege,
+  deleteUser,
+} from '../controllers/admin.controller';
+import { isAuthenticated } from '../controllers/auth.middleware';
+import { approve } from '../controllers/auth.controller';
+import 'dotenv/config';
+
+const router = express.Router();
+
+router.get('/all', isAuthenticated, isAdmin, getAllUsers);
+
+router.get('/adminstatus', isAuthenticated, isAdmin, getAllUsers);
+
+router.put('/promote', isAuthenticated, isAdmin, approve);
+
+// delete during deployment
+router.put('/autopromote', upgradePrivilege);
+
+router.delete('/:email', isAuthenticated, isAdmin, deleteUser);
+
+export default router;
